refactor: 레포지토리 객체 서비스 객체로 변경
