--- src/main/java/com/sprint/mission/discodeit/infra/UserRepository.java
@@ -1,20 +1,0 @@
-package com.sprint.mission.discodeit.infra;
-
-import com.sprint.mission.application.UserDto;
-import com.sprint.mission.application.UserRegisterDto;
-import java.util.List;
-import java.util.UUID;
-
-public interface UserRepository {
-    UserDto save(UserRegisterDto userRegisterDto);
-
-    UserDto findById(UUID id);
-
-    List<UserDto> findByName(String name);
-
-    List<UserDto> findAll();
-
-    void updateName(UUID id, String name);
-
-    void delete(UUID id);
-}
